+++ src/app/(auth)/layout.tsx
@@ -1,0 +1,12 @@
+'use client';
+import { ReactNode } from 'react'
+
+
+
+export default function Layout({ children }: { children: ReactNode }) {
+    return (
+
+        <main>{children}</main>
+
+    )
+}
