+++ src/app/(dashboard)/search/components/map/geo-albers-usa-territories.d.ts
@@ -1,0 +1,7 @@
+// geo-albers-usa-territories.d.ts
+
+declare module "geo-albers-usa-territories" {
+    // eslint-disable-next-line @typescript-eslint/no-explicit-any
+    const geoAlbersUsaTerritories: any; // Replace `any` with proper type if available.
+    export default geoAlbersUsaTerritories;
+  }
