--- src/app/home/layout.tsx
@@ -1,18 +1,0 @@
-'use client';
-
-import { Header } from "@/components/header";
-import { UserProvider } from "../userProvider";
-
-
-export default function Layout({ children }) {
-    return (
-        <UserProvider>
-
-            <Header />
-            {/* <Navbar /> */}
-            <main>{children}</main>
-            {/* <Footer /> */}
-        </UserProvider>
-
-    )
-}
