--- src/app/login/login.module.css
@@ -1,12 +1,0 @@
-.container {
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    min-height: 100vh;
-    background-color: #f0f2f5;
-  }
-  
-  .loginCard {
-    width: 100%;
-    max-width: 400px;
-  }
