--- src/interfaces.ts
@@ -1,40 +1,0 @@
-export interface Dog {
-    id: string
-    img: string
-    name: string
-    age: number
-    zip_code: string
-    breed: string
-}
-
-export interface Location {
-    zip_code: string
-    latitude: number
-    longitude: number
-    city: string
-    state: string
-    county: string
-}
-
-export interface Coordinates {
-    lat: number;
-    lon: number;
-}
-
-
-export interface Match {
-    match: string
-}
-
-
-export enum SortField {
-    BREED = 'breed',
-    NAME = 'name',
-    AGE = 'age'
-
-}
-
-export enum SortOrder {
-    ASC = 'asc',
-    DESC = 'desc'
-}
